% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\name{Class}
\alias{Class}
\title{Set an new DataClass(environment)}
\usage{
Class(class)
}
\arguments{
\item{class}{character class name}
}
\value{
enviroment(DataClass)
}
\description{
Set an new environment to store important data in a project
}
\examples{
# T_DATA <- Class()
# T_DATA$catinfo()
# T_DATA$new(T_DATA, "id", 1000, "tnbc差异相分离基因SYMOBOL列表")
# T_DATA$new(envir = T_DATA, variable = "id", value = 1000, info = "hacd")
# T_DATA$id
}
