% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{colstartlog2}
\alias{colstartlog2}
\title{Log2 dataframe start from the col}
\usage{
colstartlog2(obj, log.col = NULL)
}
\arguments{
\item{obj}{object this object must be a dataframe contain os data col and exprs data}

\item{log.col}{number the column where log2 transformate start}
}
\value{
dataframe
}
\description{
Log2 dataframe start from the col
}
\examples{
#
#        sample_id  OS  OS.time     RAB4B
# TCGA-05-4249-01A   0     1523  8.796464
# TCGA-05-4250-01A   1      121  8.482698
# in this condition, the param log.col should be set 4, meaning log2(df[,4:end])
}
