% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEeset.R
\name{DEeset}
\alias{DEeset}
\title{R6 Class DEeset : DEG analysis object}
\description{
DEG analysis pipeline object
}
\author{
Jiang
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mark}}{character the object name.}

\item{\code{eset}}{dataframe a exprs eset dataframe, \cr
col: gene, rownames: samples}

\item{\code{eset2}}{dataframe a clean dataframe from `eset` by updateGroup for different gene anlaysis\cr
col: gene, rownames: samples}

\item{\code{group}}{dataframe a phen dataframe for different gene anlaysis, \cr
the first col: Type, rownames: samples}

\item{\code{os}}{dataframe os data, rowname: samples, col: OS OS.time}

\item{\code{phen}}{dataframe phen data, rowname: samples, col: age, sex, stage ...}

\item{\code{fc.list}}{list, different fold change,  list('1.2' = log2(1.2), '1.5' = log2(1.5), '2' = log2(2), '4' = log2(4), '8' = log2(8) )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEeset-new}{\code{DEeset$new()}}
\item \href{#method-DEeset-hello}{\code{DEeset$hello()}}
\item \href{#method-DEeset-updateGroup_Eset}{\code{DEeset$updateGroup_Eset()}}
\item \href{#method-DEeset-runDEG}{\code{DEeset$runDEG()}}
\item \href{#method-DEeset-clone}{\code{DEeset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEeset-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEeset-new}{}}}
\subsection{Method \code{new()}}{
Create a new DEeset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEeset$new(mark, eset, group = NULL, os = NULL, phen = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mark}}{character the object name.}

\item{\code{eset}}{dataframe a exprs eset dataframe for different gene anlaysis, \cr
col: gene, rownames: samples}

\item{\code{group}}{dataframe a phen dataframe for different gene anlaysis, \cr
the first col: Type, rownames: samples}

\item{\code{os}}{dataframe os data, rowname: samples, col: OS OS.time}

\item{\code{phen}}{dataframe phen data, rowname: samples, col: age, sex, stage ...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DEeset` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEeset-hello"></a>}}
\if{latex}{\out{\hypertarget{method-DEeset-hello}{}}}
\subsection{Method \code{hello()}}{
test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEeset$hello()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEeset-updateGroup_Eset"></a>}}
\if{latex}{\out{\hypertarget{method-DEeset-updateGroup_Eset}{}}}
\subsection{Method \code{updateGroup_Eset()}}{
updateGroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEeset$updateGroup_Eset(
  suffix1 = "01A",
  suffix2 = "11A",
  group1 = "Tumor",
  group2 = "Normal",
  method = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suffix1}}{character the match rule1: treatment group suffix}

\item{\code{suffix2}}{character the match rule2: control group suffix}

\item{\code{group1}}{character if match true}

\item{\code{group2}}{character if match false}

\item{\code{method}}{number, default is 1, \cr
if method is not 1 will use the existed self$group to make eset2}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
no return, just update object's field group
DEG analysis pipleline
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEeset-runDEG"></a>}}
\if{latex}{\out{\hypertarget{method-DEeset-runDEG}{}}}
\subsection{Method \code{runDEG()}}{
DEG analysis pipeline, DESeq2 deg analysis, plot pca, prepare GSEA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEeset$runDEG(
  f_mark = "DE",
  outdir = "result",
  pval = 0.05,
  fdr = 0.1,
  logfc = 1,
  method = "deseq2",
  coef = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f_mark}}{character  a mark show what analysis you are doning, \cr
it will be used for subdir of outdir or filename}

\item{\code{outdir}}{character outdir}

\item{\code{pval}}{number criteria of DEG analysis: Pvalue, defaut is 0.05}

\item{\code{fdr}}{number criteria of DEG analysis: FDR/Qvalue, defaut is 0.1}

\item{\code{logfc}}{number criteria of DEG analysis: log2FC, defaut is 1}

\item{\code{method}}{character deg analysis method, default is "deseq2", can also be "edger" or "voom",\cr
if you data is microchip data, you can also use "limma"}

\item{\code{coef}}{number default is 2, do not modify this parameter unless you know what you are doing!}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list (deg analysis)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEeset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEeset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEeset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
