% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_DEres.R
\name{DEres}
\alias{DEres}
\title{R6 Class DEres : DEG analysis result object}
\description{
DEG analysis result object
}
\author{
Jiang
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resdf}}{data.frame, all gene diff analysis result}

\item{\code{glist.deg}}{character vector, the significant deg(default criterion[p<0.05,q<0.1,abs(logfc)>1])}

\item{\code{glist.deg.up}}{character vector, the significant up deg(default criterion[p<0.05,q<0.1,logfc>1])}

\item{\code{glist.deg.down}}{character vector, the significant up deg(default criterion[p<0.05,q<0.1,logfc<-1])}

\item{\code{fc.list}}{list, different fold change,  list('1.2' = log2(1.2), '1.5' = log2(1.5), '2' = log2(2), '4' = log2(4), '8' = log2(8) )}

\item{\code{glist.deg.multi}}{character vector, the significant deg by different criterion\cr
such as [p<0.01, q<0.1] logfc > log2(1.2) log2(1.5) 1 2 3 \cr
        [p<0.01, q<0.05] logfc > log2(1.2) log2(1.5) 1 2 3 \cr
        [p<0.01, q<0.01] logfc > log2(1.2) log2(1.5) 1 2 3 \cr}

\item{\code{glist.gsea}}{name list, the all gene name list(name is genename, value is the log2FC)}

\item{\code{ddslist}}{list, the diff analysis original result, the father of all_father}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DEres-new}{\code{DEres$new()}}
\item \href{#method-DEres-update.glist.deg.multi}{\code{DEres$update.glist.deg.multi()}}
\item \href{#method-DEres-pic.valcano}{\code{DEres$pic.valcano()}}
\item \href{#method-DEres-runENRICH}{\code{DEres$runENRICH()}}
\item \href{#method-DEres-clone}{\code{DEres$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEres-new"></a>}}
\if{latex}{\out{\hypertarget{method-DEres-new}{}}}
\subsection{Method \code{new()}}{
Create a new DEres object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEres$new(
  resdf,
  glist.deg,
  glist.deg.up,
  glist.deg.down,
  fc.list,
  glist.gsea,
  ddslist
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resdf}}{data.frame, all gene diff analysis result}

\item{\code{glist.deg}}{character vector, the significant deg(default criterion[p<0.05,q<0.1,abs(logfc)>1])}

\item{\code{glist.deg.up}}{character vector, the significant up deg(default criterion[p<0.05,q<0.1,logfc>1])}

\item{\code{glist.deg.down}}{character vector, the significant down deg(default criterion[p<0.05,q<0.1,logfc<-1])}

\item{\code{fc.list}}{list, different fold change,  list('1.2' = log2(1.2), '1.5' = log2(1.5), '2' = log2(2), '4' = log2(4), '8' = log2(8) )}

\item{\code{glist.gsea}}{name list, the all gene name list(name is genename, value is the log2FC)}

\item{\code{ddslist}}{list, the diff analysis original result, the father of all_father}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEres-update.glist.deg.multi"></a>}}
\if{latex}{\out{\hypertarget{method-DEres-update.glist.deg.multi}{}}}
\subsection{Method \code{update.glist.deg.multi()}}{
update field glist.deg.multi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEres$update.glist.deg.multi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEres-pic.valcano"></a>}}
\if{latex}{\out{\hypertarget{method-DEres-pic.valcano}{}}}
\subsection{Method \code{pic.valcano()}}{
plot valcano with mask gene using valcano2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEres$pic.valcano(
  filterc = "both",
  pvalue = 0.05,
  fdr = 0.1,
  logfc.p = 1,
  logfc.n = -1,
  label_geneset = NULL,
  outdir = "result",
  filename.base = "DEG_xx",
  pic.w = 7,
  pic.h = 7,
  color.not = "#BEBEBE",
  color.up = "#AC001D",
  color.down = "#3D71AD",
  ymax = NULL,
  size = 1.5,
  shape = 16,
  alpha = 0.9
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filterc}}{character the way of show picture}

\item{\code{pvalue}}{number pvalue, default 0.05}

\item{\code{fdr}}{number fdr, default 0.1}

\item{\code{logfc.p}}{number log2fc positive, default 1}

\item{\code{logfc.n}}{number log2fc negative, default -1}

\item{\code{label_geneset}}{character(n) the genelist need be shown in the valcano, default NULL}

\item{\code{outdir}}{character the picture outdir, default "result"}

\item{\code{filename.base}}{character the picture filename base, default "DEG_xx"}

\item{\code{pic.w}}{number the picture width, default 7}

\item{\code{pic.h}}{number the picture height, default 7}

\item{\code{color.not}}{character the picture color to not sig gene, default "#BEBEBE"}

\item{\code{color.up}}{character the picture color to up sig gene, default "#AC001D"}

\item{\code{color.down}}{character the picture color to down sig gene, default "#3D71AD"}

\item{\code{ymax}}{number the max ylim, default NULL}

\item{\code{size}}{number the picture point size, default 1.5 ...}

\item{\code{shape}}{number the picture point shape, default 16, or 1 ...}

\item{\code{alpha}}{number the transpate of picture, default 0.9}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEres-runENRICH"></a>}}
\if{latex}{\out{\hypertarget{method-DEres-runENRICH}{}}}
\subsection{Method \code{runENRICH()}}{
run GO analysis in batches(upgene, downgene, allgene) and write result to xlsx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEres$runENRICH(
  genelist.filter,
  outdir,
  glist.save = T,
  rungo = T,
  runkegg = T,
  rapid = T,
  orgdb = "org.Hs.eg.db",
  org_kegg = "hsa",
  sigNodes = 20
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genelist.filter}}{number(n) or missing, select which fclist to run enrich anlaysis}

\item{\code{outdir}}{character, the output dir}

\item{\code{glist.save}}{logical default is TRUE, wether save genelist}

\item{\code{rungo}}{logical default is TRUE, wether run go analysis}

\item{\code{runkegg}}{logical default is TRUE, wether run kegg analysis}

\item{\code{rapid}}{loggicla defaut is F, if set F, the function will plot go tree picture, which make cost much time}

\item{\code{orgdb}}{character deafult is "org.Hs.eg.db"}

\item{\code{org_kegg}}{character deafult is "hsa"}

\item{\code{sigNodes}}{number deafult is 20, go tree plot controls}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DEres-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DEres-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DEres$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
