% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme.add.stat}
\alias{ggtheme.add.stat}
\title{Add ggplot theme stat}
\usage{
ggtheme.add.stat(
  label.x = 1.45,
  label.y,
  label = "p.signif",
  method,
  comparisons,
  ref.group
)
}
\arguments{
\item{label.x}{numner the stat's x}

\item{label.y}{numner the stat's y, for example, max(p.data[, taget]), can aslo be a numeric(n) as c(29, 35, 40) for multi compare}

\item{label}{character the stat's show method, such as "p.signif", "p.format", ...}

\item{method}{character stat method, such as "t.test", "wilcox.test", "anova", "kruskal.test"}

\item{comparisons}{list multi compare group, list(c("0.5","1"), c("1", "2"),c("0.5", "2"))}

\item{ref.group}{character the ref group}
}
\value{
ggtheme
}
\description{
Add ggplot theme stat
}
\examples{
NA
}
\author{
Jiang
}
