% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme.update.text.legend}
\alias{ggtheme.update.text.legend}
\title{Update ggplot theme text legend}
\usage{
ggtheme.update.text.legend(size = 8, color = "black", family = "serif")
}
\arguments{
\item{size}{number font size, default is 8}

\item{color}{character font color, default is "black", can be rgb value, such as "#00FFFF"}

\item{family}{character font family, default is "serif"}
}
\value{
ggtheme
}
\description{
Update ggplot theme text
}
\author{
Jiang
}
