% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme.update.text}
\alias{ggtheme.update.text}
\title{Update ggplot theme text}
\usage{
ggtheme.update.text(size = 14, color = "black", family = "serif")
}
\arguments{
\item{size}{number font size, default is 14}

\item{color}{character font color, default is "black", can be rgb value, such as "#00FFFF"}

\item{family}{character font family, default is "serif"}
}
\value{
ggtheme
}
\description{
Update ggplot theme text
}
\author{
Jiang
}
