% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\name{lzdelete}
\alias{lzdelete}
\title{Delete a attribute to an environment}
\usage{
lzdelete(envir, variable)
}
\arguments{
\item{envir}{object the DataClass name(environment name)}

\item{variable}{character the attribute name which you want to delete}
}
\value{
Null
}
\description{
Delete a attribute and its information in an environment
}
\examples{
# lzdelete(envir = test, variable = "id")
}
\author{
Jiang
}
