% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObject.R
\name{lznew}
\alias{lznew}
\title{Add a attribute to an environment}
\usage{
lznew(envir, variable, value, info = NA)
}
\arguments{
\item{envir}{object the DataClass name(environment name)}

\item{variable}{character new attribute name}

\item{value}{object new attribute value}

\item{info}{character the information of new attribute}
}
\value{
environment's new attribute
}
\description{
Add a attribute to an environment, in this package, it's usually used to add a attribute/variable to the DataClass Object(environment)
}
\examples{
# lznew(envir = test, variable = "id", value = 1000, info = "Hellow")
}
\author{
Jiang
}
