% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{LZplot.dens}
\alias{LZplot.dens}
\title{plot density}
\usage{
LZplot.dens(data, col)
}
\arguments{
\item{data}{dataframe expr dataframe}

\item{col}{character the taget gene name}
}
\value{
density plot
}
\description{
plot density
}
\examples{
#
}
