% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{lzsurvkm}
\alias{lzsurvkm}
\title{Xena surv pipeline}
\usage{
lzsurvkm(eset_os, tagetGene, method)
}
\arguments{
\item{eset_os}{data.frame a dataframe contain os info and exprs(row is sample)}

\item{tagetGene}{character genes}

\item{method}{character cut point}
}
\value{
plot
}
\description{
km surv analysis, km plot + density
}
\examples{
#
}
