% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{xena.surv.cut}
\alias{xena.surv.cut}
\title{xena.surv.cut}
\usage{
xena.surv.cut(eset_os.df, tagetGene = "SIN3B", method = 2)
}
\arguments{
\item{eset_os.df}{dataframe}

\item{tagetGene}{character}

\item{method}{character}
}
\value{
list
}
\description{
xena.surv.cut
}
