% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\name{xena.surv.getPvale}
\alias{xena.surv.getPvale}
\title{xena.surv.getPvale}
\usage{
xena.surv.getPvale(rt, num.tran = 365, main.text = "")
}
\arguments{
\item{rt}{dataframe}

\item{num.tran}{number}

\item{main.text}{character}
}
\value{
list
}
\description{
xena.surv.getPvale
}
