% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xena_flow.R
\name{XENADB.init}
\alias{XENADB.init}
\title{Create XENA OBJECT}
\usage{
XENADB.init(
  eset.count.file,
  eset.fpkm.file,
  os.file,
  phen.file,
  eset.count.info = "xena原始count数据(log2(x+1))",
  eset.fpkm.info = "xena原始fpkm数据(log2(x+1))",
  os.info = "xena原始OS数据",
  phen.info = "xena原始phen数据",
  class = "XENA"
)
}
\arguments{
\item{eset.count.file}{character the file dir}

\item{eset.fpkm.file}{character the file dir}

\item{os.file}{character the file dir}

\item{phen.file}{character the file dir}

\item{eset.count.info}{character the information}

\item{eset.fpkm.info}{character the information}

\item{os.info}{character the information}

\item{phen.info}{character the information}

\item{class}{character do not modifiy the param in xena flow}
}
\value{
`XENA` OBJECT
}
\description{
The original xena data object.
}
\author{
Jiang
}
